feature/navigation-order-details

+++ user/src/main/java/org/the_chance/honeymart/ui/feature/orders/OrderUiEffect.kt
@@ -2,7 +2,7 @@ package org.the_chance.honeymart.ui.feature.orders
 
 
 sealed class OrderUiEffect {
-    data class ClickOrderEffect(val orderId: Int) : OrderUiEffect()
+    data class ClickOrderEffect(val orderId: Long) : OrderUiEffect()
     object UnAuthorizedUserEffect : OrderUiEffect()
     object ClickDiscoverMarketsEffect : OrderUiEffect()
 }
